feat[audio.c]: added getter functions
The following getter functions were added:
- struct autx *get_autx()
- struct audio_recv *get_aurx()
- uint32_t get_first_timestamp()
- uint64_t get_rtp_time_out()
- uint32_t get_packet_time()

These are used to get the corresponding struct members for
debugging and research purposes

refactor[audio.c; in function encode_rtp_send]: commented code out

The code between both '//---' is not not used, but I left it there
in case it  might be needed later.

refactor[audio.c; in function audio_debug]: commented code out and added
code to get metadata

The reasobn, why parts of the code were commented out, was to make the
console output  less verbose.

The metadata consists of information such as latency, the size of the audio
jitter buffer, etc. These are then piped from stderr to a file which is then
cleaned up and analysed later via python.